+++ CHANGELOG.md
@@ -4,7 +4,13 @@ This changelog follows the specifications detailed in: [Keep a Changelog](https:
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html), although we have not yet reached a `1.0.0` release.
 
 
-## Version 0.7.0 - Unreleased
+## Version 0.7.1 - Unreleased
+
+### Changed
+* input to `ensure_rng` can now be a string.
+
+
+## Version 0.7.0 - Released 2024-08-14
 
 ### Removed
 * Removed support for Python 3.6 and 3.7